feat: prepare sql-formatter to format SQL files
As we add more SQL files to the repo it makes sense to enforce a common
format for them. This commit adds a new tool to `treefmt.nix` which does
exactly that, `sql-formatter`. Furthermore, all existing `*.sql` files
in the repo are formatted accordingly.

The formatter is not yet active in the `treefmt.nix`, due to a bug [1],
which causes the CI to fail on every invocation of `nix fmt -- --ci`

[1] https://github.com/numtide/treefmt-nix/issues/400

Signed-off-by: wucke13 <wucke13+github@gmail.com>